fixed bug sample extraction
